fix: fixed messed up migrations + added url conf for subscription app
